+++ frontend/src/utils/ColorPalette.js
@@ -1,3 +1,24 @@
-const colorPalette = ["red", "yellow", "orange", "blue", "purple"];
+const colorPalette = [
+  "red",
+  "crimson",
+  "firebrick",
+  "darkred",
+  "orange",
+  "darkorange",
+  "gold",
+  "yellow",
+  "blue",
+  "royalblue",
+  "navy",
+  "indigo",
+  "purple",
+  "rebeccapurple",
+  "magenta",
+  "hotpink",
+  "deeppink",
+  "pink",
+  "brown",
+  "saddlebrown",
+];
 
 export default colorPalette;