+++ scripts/process_similarity.py
@@ -130,4 +130,3 @@ def process_similarity(
             json.dump(similarity, outfile, indent=2)
 
     return similarity
-